# Conflicts:
#	longjian-measure-app/src/main/java/com/longfor/longjian/measure/app/appService/CheckItemMeasureService/OapiCheckItemMeasureServiceImpl.java

+++ longjian-measure-app/src/main/java/com/longfor/longjian/measure/app/vo/appMeasureSyncVo/ResultListVo.java
@@ -61,11 +61,11 @@ public class ResultListVo {
     /**
      * 更新时间
      */
-    private Long update_at;
+    private Integer update_at;
     /**
      * 删除时间(0表示未删除)
      */
-    private Long delete_at;
+    private Integer delete_at;
     /**
      * 描区uuid
      */
